Keep domain/mod.rs and app_input/mod.rs under 1000-line CI gate (Fixes #121)
Extract domain inline test module to src/domain/tests.rs and the runtime-binding helpers to src/app_input/agent_runtime.rs. Behavior-preserving; no logic changes; test count unchanged.
